@@ -2,7 +2,7 @@ package com.quizplanner.quizPlanner.player
 
 import android.os.Handler
 import android.os.Looper
-import android.support.annotation.RestrictTo
+import androidx.annotation.RestrictTo
 
 import android.text.TextUtils
 import android.webkit.JavascriptInterface
