@@ -1,12 +1,12 @@
 package com.quizplanner.quizPlanner.player.ui.menu.defaultMenu
 
 import android.content.Context
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.RecyclerView
+import androidx.core.content.ContextCompat
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import com.quizplanner.quizPlanner.player.ui.menu.MenuItem
 import com.quizplanner.quizPlanner.R
 
