@@ -5,10 +5,10 @@ import android.util.AttributeSet
 import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
-import android.support.annotation.LayoutRes
-import android.arch.lifecycle.Lifecycle
-import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.OnLifecycleEvent
+import androidx.annotation.LayoutRes
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.OnLifecycleEvent
 import com.quizplanner.quizPlanner.R
 import com.quizplanner.quizPlanner.player.YouTubePlayer
 import com.quizplanner.quizPlanner.player.listeners.AbstractYouTubePlayerListener
@@ -16,16 +16,16 @@ import com.quizplanner.quizPlanner.player.listeners.YouTubePlayerCallback
 import com.quizplanner.quizPlanner.player.listeners.YouTubePlayerFullScreenListener
 import com.quizplanner.quizPlanner.player.listeners.YouTubePlayerListener
 import com.quizplanner.quizPlanner.player.options.IFramePlayerOptions
-import com.quizplanner.quizPlanner.player.utils.FullScreenHelper
 import com.quizplanner.quizPlanner.player.ui.PlayerUiController
+import com.quizplanner.quizPlanner.player.utils.FullScreenHelper
 import com.quizplanner.quizPlanner.player.utils.loadOrCueVideo
 
 
-class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0):
+class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0) :
         SixteenByNineFrameLayout(context, attrs, defStyleAttr), LifecycleObserver {
 
-    constructor(context: Context): this(context, null, 0)
-    constructor(context: Context, attrs: AttributeSet? = null): this(context, attrs, 0)
+    constructor(context: Context) : this(context, null, 0)
+    constructor(context: Context, attrs: AttributeSet? = null) : this(context, attrs, 0)
 
     private val legacyTubePlayerView: LegacyYouTubePlayerView = LegacyYouTubePlayerView(context)
     private val fullScreenHelper = FullScreenHelper(this)
@@ -52,16 +52,16 @@ class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleA
 
         typedArray.recycle()
 
-        if(!enableAutomaticInitialization && useWebUi) {
+        if (!enableAutomaticInitialization && useWebUi) {
             throw IllegalStateException("YouTubePlayerView: 'enableAutomaticInitialization' is false and 'useWebUi' is set to true. " +
                     "This is not possible, if you want to manually initialize YouTubePlayerView and use the web ui, " +
                     "you should manually initialize the YouTubePlayerView using 'initializeWithWebUi'")
         }
 
-        if(videoId == null && autoPlay)
+        if (videoId == null && autoPlay)
             throw IllegalStateException("YouTubePlayerView: videoId is not set but autoPlay is set to true. This combination is not possible.")
 
-        if(!useWebUi) {
+        if (!useWebUi) {
             legacyTubePlayerView.getPlayerUiController()
                     .enableLiveVideoUi(enableLiveVideoUi)
                     .showYouTubeButton(showYouTubeButton)
@@ -81,8 +81,8 @@ class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleA
             }
         }
 
-        if(enableAutomaticInitialization) {
-            if(useWebUi) legacyTubePlayerView.initializeWithWebUi(youTubePlayerListener, handleNetworkEvents)
+        if (enableAutomaticInitialization) {
+            if (useWebUi) legacyTubePlayerView.initializeWithWebUi(youTubePlayerListener, handleNetworkEvents)
             else legacyTubePlayerView.initialize(youTubePlayerListener, handleNetworkEvents)
         }
 
@@ -118,7 +118,7 @@ class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleA
      * @see YouTubePlayerView.initialize
      */
     fun initialize(youTubePlayerListener: YouTubePlayerListener, handleNetworkEvents: Boolean) {
-        if(enableAutomaticInitialization) throw IllegalStateException("YouTubePlayerView: If you want to initialize this view manually, you need to set 'enableAutomaticInitialization' to false")
+        if (enableAutomaticInitialization) throw IllegalStateException("YouTubePlayerView: If you want to initialize this view manually, you need to set 'enableAutomaticInitialization' to false")
         else legacyTubePlayerView.initialize(youTubePlayerListener, handleNetworkEvents, null)
     }
 
@@ -140,7 +140,7 @@ class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleA
      * @see YouTubePlayerView.initialize
      */
     fun initializeWithWebUi(youTubePlayerListener: YouTubePlayerListener, handleNetworkEvents: Boolean) {
-        if(enableAutomaticInitialization) throw IllegalStateException("YouTubePlayerView: If you want to initialize this view manually, you need to set 'enableAutomaticInitialization' to false")
+        if (enableAutomaticInitialization) throw IllegalStateException("YouTubePlayerView: If you want to initialize this view manually, you need to set 'enableAutomaticInitialization' to false")
         else legacyTubePlayerView.initializeWithWebUi(youTubePlayerListener, handleNetworkEvents)
     }
 
@@ -150,7 +150,7 @@ class YouTubePlayerView(context: Context, attrs: AttributeSet? = null, defStyleA
      * This function is called only once.
      */
     fun getYouTubePlayerWhenReady(youTubePlayerCallback: YouTubePlayerCallback) =
-        legacyTubePlayerView.getYouTubePlayerWhenReady(youTubePlayerCallback)
+            legacyTubePlayerView.getYouTubePlayerWhenReady(youTubePlayerCallback)
 
     /**
      * Use this method to replace the default Ui of the player with a custom Ui.
