@@ -1,13 +1,14 @@
 package com.quizplanner.quizPlanner.exchange
 
 import android.content.Context
+import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import com.quizplanner.quizPlanner.R
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
-import rx.Observable
+import io.reactivex.Observable
 import javax.net.ssl.HttpsURLConnection
 
 
@@ -66,7 +67,7 @@ class RetrofitService private constructor(context: Context) {
         mRetrofit = Retrofit.Builder()
                 .baseUrl(Urls.getApiBaseUrl(context))
                 .addConverterFactory(GsonConverterFactory.create())
-                .addCallAdapterFactory(RxJavaCallAdapterFactory.create())
+                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
                 .client(client)
                 .build()
     }
