@@ -1,13 +1,13 @@
 package com.quizplanner.quizPlanner.ui
 
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.RecyclerView
 import com.quizplanner.quizPlanner.QuizPlanner
 import com.quizplanner.quizPlanner.R
 import com.quizplanner.quizPlanner.model.Quiz
